# backend/build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Todo": {
				"properties": {
					"completed": {
						"type": "boolean"
					},
					"note": {
						"type": "string",
						"nullable": true
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"completed",
					"note",
					"title",
					"id"
				],
				"type": "object"
			},
			"Task": {
				"properties": {
					"completed": {
						"type": "boolean"
					},
					"note": {
						"type": "string",
						"nullable": true
					},
					"title": {
						"type": "string"
					},
					"todoId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"completed",
					"note",
					"title",
					"todoId",
					"id"
				],
				"type": "object"
			},
			"TodoAndTask": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Todo"
					},
					{
						"properties": {
							"tasks": {
								"items": {
									"$ref": "#/components/schemas/Task"
								},
								"type": "array"
							}
						},
						"required": [
							"tasks"
						],
						"type": "object"
					}
				]
			},
			"TodoPostParams": {
				"properties": {
					"note": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"title"
				],
				"type": "object"
			},
			"Partial_TodoPostParams_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"note": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"TodoPutParams": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Partial_TodoPostParams_"
					},
					{
						"properties": {
							"completed": {
								"type": "boolean"
							}
						},
						"type": "object"
					}
				]
			},
			"Pick_AddTaskParams.Exclude_keyofAddTaskParams.todoId__": {
				"properties": {
					"title": {
						"type": "string"
					},
					"note": {
						"type": "string"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_AddTaskParams.todoId_": {
				"$ref": "#/components/schemas/Pick_AddTaskParams.Exclude_keyofAddTaskParams.todoId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"TaskPostParams": {
				"$ref": "#/components/schemas/Omit_AddTaskParams.todoId_"
			},
			"Pick_UpdateTaskParams.Exclude_keyofUpdateTaskParams.id__": {
				"properties": {
					"title": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"completed": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_UpdateTaskParams.id_": {
				"$ref": "#/components/schemas/Pick_UpdateTaskParams.Exclude_keyofUpdateTaskParams.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"TaskPutParams": {
				"$ref": "#/components/schemas/Omit_UpdateTaskParams.id_"
			},
			"User": {
				"properties": {
					"name": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"createdAt",
					"id"
				],
				"type": "object",
				"description": "Model User"
			},
			"UserPostParams": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/todo": {
			"get": {
				"operationId": "GetAllTodos",
				"responses": {
					"200": {
						"description": "ToDo配列",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TodoAndTask"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "全ToDoを取得する",
				"summary": "全ToDo取得",
				"tags": [
					"Todo"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTodo",
				"responses": {
					"200": {
						"description": "登録したToDo",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Todo"
								}
							}
						}
					}
				},
				"description": "ToDo登録",
				"summary": "ToDo登録",
				"tags": [
					"Todo"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TodoPostParams"
							}
						}
					}
				}
			}
		},
		"/todo/{todoId}": {
			"put": {
				"operationId": "UpdateTodo",
				"responses": {
					"200": {
						"description": "登録したToDo",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Todo"
								}
							}
						}
					}
				},
				"description": "ToDo更新",
				"summary": "ToDo更新",
				"tags": [
					"Todo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "todoId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TodoPutParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTodo",
				"responses": {
					"200": {
						"description": "削除したToDo",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Todo"
								}
							}
						}
					}
				},
				"description": "ToDo削除",
				"summary": "ToDo削除",
				"tags": [
					"Todo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "todoId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/todo/{todoId}/task": {
			"post": {
				"operationId": "AddTask",
				"responses": {
					"200": {
						"description": "登録したタスク",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Todo"
								}
							}
						}
					}
				},
				"description": "タスクの登録",
				"summary": "指定のToDoにタスクを登録する",
				"tags": [
					"Todo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "todoId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TaskPostParams"
							}
						}
					}
				}
			}
		},
		"/todo/task/{taskId}": {
			"put": {
				"operationId": "UpdateTask",
				"responses": {
					"200": {
						"description": "更新したタスク",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Todo"
								}
							}
						}
					}
				},
				"description": "タスクの更新",
				"summary": "タスクを更新する",
				"tags": [
					"Todo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TaskPutParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTask",
				"responses": {
					"200": {
						"description": "更新したタスク",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Todo"
								}
							}
						}
					}
				},
				"description": "タスクの削除",
				"summary": "タスクを削除する",
				"tags": [
					"Todo"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "ユーザ配列",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "全ユーザを取得す",
				"summary": "全ユーザ取得",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "登録ユーザ",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"description": "ユーザ登録",
				"summary": "ユーザ登録",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserPostParams"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "ユーザ",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					},
					"404": {
						"description": "データが取得できなかった"
					}
				},
				"description": "UserIDを指定してユーザを取得",
				"summary": "ユーザ取得",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "ユーザID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}